Update index.html
Hi Marius,

I hope you are doing well!

We need to update the website and I thought it's a opportunity start to learn how to use GitHub. So that I can make changes on the website without you completely losing the oversight over the code, just in case we need to ask you once again to fix the mess I potentially create. <3

Lots of good wishes,
Moritz